+++ src/placers/weights-stair.js
