+++ src/placers/weights-uniform.js
